e2e test for error response
