+++ src/applications/personalization/rated-disabilities/tests/e2e/ratedDisabilitiesError.e2e.spec.js
@@ -1,0 +1,40 @@
+const E2eHelpers = require('platform/testing/e2e/helpers');
+const Timeouts = require('platform/testing/e2e/timeouts');
+const Auth = require('platform/testing/e2e/auth');
+const Mock = require('platform/testing/e2e/mock-helpers');
+const ErrorMockData = require('../mockdata/error-response.json');
+
+function runRatedDisabilitiesTest(browser) {
+  browser.assert.containsText(
+    '.usa-alert-heading',
+    'No disability rating found',
+  );
+  browser.assert.visible('.usa-alert-text');
+  browser.assert.containsText(
+    '.usa-alert-text p:nth-of-type(1)',
+    "We sorry. We can't find a disability rating matched with the name, date of birth, and social secuity number you provided in our Veteran records.",
+  );
+}
+
+function generateErrorData(token) {
+  return Mock(token, ErrorMockData);
+}
+
+function begin(browser) {
+  browser.perform(done => {
+    const token = Auth.getUserToken();
+    generateErrorData(token).then(() => {
+      Auth.logIn(
+        token,
+        browser,
+        '/disability/my-rated-disabilities',
+        3,
+      ).waitForElementVisible('.usa-alert-heading', Timeouts.slow);
+    });
+    browser.pause(Timeouts.slow);
+    runRatedDisabilitiesTest(browser);
+    done();
+  });
+}
+
+module.exports = E2eHelpers.createE2eTest(begin);
