+++ src/applications/personalization/rated-disabilities/tests/mockdata/200-response.json
@@ -1,0 +1,46 @@
+{
+  "path": "/v0/disability_compensation_form/rated_disabilities",
+  "verb": "get",
+  "value": {
+    "data": {
+      "attributes": {
+        "ratedDisabilities": [
+          {
+            "decisionCode": "SVCCONNECTED",
+            "decisionText": "Service Connected",
+            "diagnosticCode": "5238",
+            "effectiveDate": "2008-10-01T05:00:00.000+00:00",
+            "name": "Diabetes mellitus0",
+            "ratedDisabilityId": "0",
+            "ratingDecisionId": "63655",
+            "ratingPercentage": "100",
+            "relatedDisabilityDate": "2012-03-09T21:22:09.000+00:00",
+            "specialIssues": [
+              {
+                "code": "TRM",
+                "name": "Personal Trauma PTSD"
+              }
+            ]
+          },
+          {
+            "decisionCode": "SVCCONNECTED",
+            "decisionText": "Service Connected",
+            "diagnosticCode": "5238",
+            "effectiveDate": "2008-10-01T05:00:00.000+00:00",
+            "name": "Diabetes mellitus0",
+            "ratedDisabilityId": "0",
+            "ratingDecisionId": "63655",
+            "ratingPercentage": "100",
+            "relatedDisabilityDate": "2012-03-09T21:22:09.000+00:00",
+            "specialIssues": [
+              {
+                "code": "TRM",
+                "name": "Personal Trauma PTSD"
+              }
+            ]
+          }
+        ]
+      }
+    }
+  }
+}
