+++ src/applications/personalization/rated-disabilities/tests/mockdata/error-response.json
@@ -1,0 +1,18 @@
+{
+  "path": "/v0/disability_compensation_form/rated_disabilities",
+  "verb": "get",
+  "value": {
+    "data": {
+      "attributes": {
+        "errors": [
+          {
+            "title": "error",
+            "detail": "access denied",
+            "code": "404",
+            "status": "404"
+          }
+        ]
+      }
+    }
+  }
+}
